Added install instructions
