+++ README.md
@@ -1,1 +1,11 @@
 # ZumoRobot
+
+## Dependencies
+In order to compile and run this program, several **dependencies**(libraries) has to be installed.
+These can be found under "dependencies".
+
+### Install instructions
+Each library has to be copied to the local Arduino-installation "library"-folder, inside sketchbook location. To see the sketchbook folder-location, go to **preferences** and "Sketchbook location" in the Arduino-IDE.
+
+Alternativly, see
+[How to install libraries](https://www.arduino.cc/en/Guide/Libraries)
