+++ dependencies/libraries/LSM303/LICENSE.txt
