+++ dependencies/libraries/LSM303/README.md
