+++ dependencies/libraries/LSM303/examples/Heading/Heading.ino
