+++ dependencies/libraries/LSM303/library.properties
