+++ dependencies/libraries/QTRSensors/README.textile
